Prevent empty filepaths from being cleaned to .
The function which cleans and expands filepaths for config options
erroneously converted the empty string to a single dot.  Correct this
by checking for an empty path as input and returning the empty path
back.

Discovered and previously fixed in dcrd, which contained a ported
version of dcrwallet's filepath cleaning.