Temporarily disable gofmt on CI
The formatting rules have changed for Go 1.11, and with the release so
close several developers are already testing the betas.  To prevent
gofmt formatting from being a blocking issue when submitting pull
requests, gofmt checks on Travis-CI are being temporarily disabled.

Gofmt checks will be reenabled and using Go 1.11 formatting rules when
Go 1.11 is released.