rpc: Add SpvSync request (#1221)
This request allows a wallet set to --nointialload to use the recently
added spv functionality.  A long-lived response stream will be used
to return various notifications about the state of the wallet syncing.
For this initial version, we simply provide when the wallet is synced
or unsynced.

Users also may choose to set the spv_connect option to use persitent
peers instead of relying on DNS seeders to finds peers.

Also, there is a discover_accounts boolean (which requires
private_passphrase to be set), that will search for any used accounts,
which is typically done during restoring from seed.