Deprecate and hide --automaticrepair
This option has not done anything for quite some time, but there was confusion
about being able to use it to as it showed up in the config.  Hide it and the
--rollbacktest options from appearing in the help output, but do not remove them
entirely from the config to prevent parsing errors if the options are set in old
config files.  Print warnings if either is specified.