Improve various aspects of run_tests.sh.
The source tree is now recursively copied with cp rather than rsync.
Sources are copied outside of any GOPATH which removes the need to set
the GO111MODULE=on environment variable.

The :Z attribute has been added to the -v mount option.  This applies
the correct SELinux settings so the container is able to access the
files inside /src.  Tested on Fedora 28 using the default SELinux
policies.  See [1] for further details about this attribute.

Support for podman (a drop-in replacement for docker) is also added.
To use, call the script as './run_tests.sh podman'.

While here, also remove 'go mod verify' from the test script.  Module
verification is implicitly done for any build/test operation, so this
is unnecessary and running it again marginally slows down CI testing.

[1] https://www.projectatomic.io/blog/2015/06/using-volumes-with-docker-can-cause-problems-with-selinux/