Allow process interrupt when blocked on DB lock.
This change moves the opening of an existing wallet DB into a
background goroutine so that main can return and the process will exit
if an interrupt signal is sent to a wallet process that is blocked on
a database lock.

To support this change, the loader mutex must be released during the
loader's open database call.