Watch gap limit addresses after last returned address
This changes the behavior of the code which watches future address
usage by modifying it to watch gap limit number of addresses after the
last derived or returned address, rather than past the last address
which is marked used.  This fixes address watching issues on heavily
used wallets which derive many addresses by ensuring that all
addresses derived for an operation are part of the watched address
set.  Before this change, it was possible that addresses were returned
past the gap limit of the last address marked used, which resulted in
these addresses not being watched and mined transactions involving the
address to be missed.