Add CoinShuffle++ support
This commit introduces support for creating CoinJoin transactions for purchasing
split transactions in ticketbuying, as well as mixing of individual mix change
outputs into smaller standard denominations.

At the moment, for ticket buying, this process is only available through the
automated ticket buyer.  Future commits will enable this feature through the
ticketbuying RPCs.

Note: this work is being published for early adopters.  Some features (such as
buying tickets for a VSP) are intentionally removed in this early version.
These features will be reintroduced before this change is merged to master, but
they are are not currently compatible with the requirements to perform
CoinShuffle++ and will not be enabled until all VSP address reusage can be
eliminated.

A testnet server is available at cspp.decred.org:15760.