Remove additional legacy references to JSON-RPC server
The package implementing the server was already renamed to reflect the
decision that the server is neither legacy nor unmaintained.  The only
remaining references to the server being legacy are in dcrwallet
command line and config flags, which can't be changed without
deprecating them in at least one release cycle.