Fix and simplify error detection using errors.As
In addition to simplification and correctly handling wrapped errors, a
couple instances of detecting for a wrapped dcrjson.RPCError were
corrected.  This is not the same error type that wsrpc returns, and
the dcrd RPC client already ignores duplicate transaction errors when
publishing through sendrawtransaction, so there is no need to repeat
this behavior after each client call.