Partition watched outpoints into multiple LoadTxFilter calls
It is a bad idea to try and register all watched outpoints from a
single call, due to the network backend possibly being a dcrd instance
and JSON-RPC being used.  Partition the data into sets of 64 outpoints
each and perform multiple calls instead of registering all of them
single-shot.

This will reduce the likelihood that a send timeout is hit on large
data sets.