Avoid child index cursor backwards resets
Some operations delay the database update recording the last returned
address of an account branch.  This may cause the lookahead address
watching code to set a new cursor which does not reflect these new
returned addresses, effectively causing the next child index for a new
address in the account to be reset and go backwards.  This corrects
this bug by considering the current difference between the in-memory
child index management and the values read from the db.