Silence ticketbuyer insufficient balance errors
There was a bug in the detection of insufficient balance errors
(reversed parameters passed to errors.Is) that was causing these
errors to be logged when they were intended to be silenced.  Since the
calling code already logs errors (and correctly silences the logs for
insufficient balance errors), just remove this unneeded code.