Use source account as change account for gRPC ticketbuyer config
A breaking change to the ticketbuyer module introduced a ChangeAccount
field to the config.  This was a breaking change due to the field now
being required to keep previous behavior, otherwise the change account
would default to the zero account.  Specify this account explicitly in
the gRPC RunTicketBuyer method handler, so that the previous behavior
is retained.