Avoid address cursor resetting during address discovery
This is a similar fix as 848d33040d, where the address cursor was
occasionally being reset.  In this case, the cause was occuring during
address discovery, and it was possible to introduce address reusage
this way when the RPC syncing process temporarily disconnected and
reconnected.

Tests have been added to ensure the the reset fix works as intended.
These tests fail without the changes in discovery.go.

A few minor changes to the wallet package's testing features were
added.  A peerFuncs type was introduced to allow definitions of custom
implementations of the Peer interface.  Wallet logs may also be viewed
now when running tests, though due to the package global logger, it is
recommended to only use this feature when running one test at a time
to avoid intertwined logs.  For example, to use it with the new test:

  $ go test -run TestDiscoveryCursorPos -v -args -log