Fix getaddressesbyaccount for the imported account
An ImportedAddresses method is added to Wallet which returns
[]KnownAddress for each imported address of an account.  The method
takes an account name, but currently the only valid account name is
"imported".

Some infractructure was added to the jsonrpc package to simplify the
marshaling of address arrays and reduce memory usage.  This can be
reused for other methods later, to avoid the need to allocate []string
objects for JSON serialization.