rpc: Fix grpc binding generation
The addition of the go_package directive in commit 3710c8d8 broke
regenerating the scripts locally.

This fixes the issue by using the source_relative option when generating
bindings.

It also fixes the docker version by updating to the latest golang docker
image and mounting from the root wallet source dir so that module
directives are correctly accounted for inside the image.