rpc: Add GetRawBlocks and GetCFilters gRPC methods
This adds the GetCFilters method to the WalletService gRPC service. This
allows clients to fetch the wallet-stored and header-validated version 2
committed filters so that non-wallet-related transactions can be
searched for.

It also adds a new service (NetworkService) with a GetRawBlock call. The
service is intended to be used for calls which fetch or push data that
might not be backed by the wallet into the P2P network.

This particular method is useful for external callers to fetch a full
block after they have determined it (probably) contains transactions
they are interested in by checking the respective cfilter.