Introduce addtransaction JSON-RPC, remove unnecessary others
This adds a new JSON-RPC method to manually add a transaction to the
wallet and record it mined in a block.  This is particularly useful
for the upcoming VSP design which will require spending transaction
outputs from transactions which have already been mined by the time a
private key was imported.  Though it is possible to record these
transaction using a rescan, this is more work for voting wallets which
should avoid long database updates as much as possible.

The addticket and stakepooluserinfo RPCs are removed in this change.
addticket usage can be replaced with addtransaction.
stakepooluserinfo is no longer used by stakepoold to decide what
tickets shall be voted for, and is therefore no longer needed.  Any
remaining use of stakepooluserinfo by dcrstakepool should be replaced
by stakepoold itself providing the hashes of tickets it will vote
with.