Unlock previous outpoints during transaction processing
When a transaction is processed (whether it is mined or in mempool),
any referenced previous outputs should be unlocked by the wallet.
Transaction creation already unlocked selected outputs when
transactions were created by the same wallet, but transactions seen
from the network were not causing spent locked outpoints to be
removed.