Fix lock ordering
This change fixes a lock ordering issue between acquiring the address
and db locks.  After this change, the order of acquired locks when
deriving the voting and subsidy addresses when buying tickets is to
first acquire the address mutex, and then the db lock.  This is
accomplished by moving the address returning functions outside of a db
update.

Since the other order (acquiring the db lock first, and then the
address lock) is incorrect, the addrFunc signature is also changed to
remove the now unused maybeDBTX parameter, and this func will always
call persistReturnedChild with a nil dbtx so that a transaction is
only opened with the address lock held.